chore: #9 패키지 구조 변경
domain package를 생성해서 기존 global과 도메인 패키지들을 구분하였다..
