--- src/main/java/site/board/boardtraining/article/constant/ArticleCommentStatus.java
@@ -1,5 +1,0 @@
-package site.board.boardtraining.article.constant;
-
-public enum ArticleCommentStatus {
-    PUBLIC, SECRET
-}
