--- src/main/java/site/board/boardtraining/article/constant/ArticleStatus.java
@@ -1,5 +1,0 @@
-package site.board.boardtraining.article.constant;
-
-public enum ArticleStatus {
-    OPEN, HIDE
-}
