--- src/main/java/site/board/boardtraining/article/entity/ArticleComment.java
@@ -1,79 +1,0 @@
-package site.board.boardtraining.article.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import site.board.boardtraining.article.constant.ArticleCommentStatus;
-import site.board.boardtraining.global.audit.BaseEntity;
-import site.board.boardtraining.member.entity.Member;
-
-import java.util.Objects;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(
-        name = "article_comment",
-        indexes = {
-                @Index(columnList = "createdAt")
-        }
-)
-@Entity
-public class ArticleComment
-        extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 10000, nullable = false)
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    @Column(length = 50, nullable = false)
-    private ArticleCommentStatus status;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Article article;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Member member;
-
-    private ArticleComment(
-            String content,
-            ArticleCommentStatus status,
-            Article article,
-            Member member
-    ) {
-        this.content = content;
-        this.status = status;
-        this.article = article;
-        this.member = member;
-    }
-
-    public static ArticleComment of(
-            String content,
-            ArticleCommentStatus status,
-            Article article,
-            Member member
-    ) {
-        return new ArticleComment(
-                content,
-                status,
-                article,
-                member
-        );
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof ArticleComment that)) return false;
-        return id != null && id.equals(that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
