--- src/main/java/site/board/boardtraining/auth/dto/LoginRequest.java
@@ -1,6 +1,0 @@
-package site.board.boardtraining.auth.dto;
-
-public record LoginRequest(
-        String personalId,
-        String password
-) {}
