--- src/main/java/site/board/boardtraining/auth/handler/LoginSuccessHandler.java
@@ -1,31 +1,0 @@
-package site.board.boardtraining.auth.handler;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.DefaultRedirectStrategy;
-import org.springframework.security.web.RedirectStrategy;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Slf4j
-@Component
-public class LoginSuccessHandler implements AuthenticationSuccessHandler {
-
-    private RedirectStrategy redirectStrategy = new DefaultRedirectStrategy();
-
-    @Override
-    public void onAuthenticationSuccess(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            Authentication authentication
-    ) throws IOException, ServletException {
-        log.info("LoginSuccessHandler.onAuthenticationSuccess");
-
-        redirectStrategy.sendRedirect(request, response, "/api/underground");
-    }
-}
