+++ src/main/java/site/board/boardtraining/domain/article/constant/ArticleCommentStatus.java
@@ -1,0 +1,5 @@
+package site.board.boardtraining.domain.article.constant;
+
+public enum ArticleCommentStatus {
+    PUBLIC, SECRET
+}
