+++ src/main/java/site/board/boardtraining/domain/article/constant/ArticleStatus.java
@@ -1,0 +1,5 @@
+package site.board.boardtraining.domain.article.constant;
+
+public enum ArticleStatus {
+    OPEN, HIDE
+}
