+++ src/main/java/site/board/boardtraining/domain/auth/dto/LoginRequest.java
@@ -1,0 +1,6 @@
+package site.board.boardtraining.domain.auth.dto;
+
+public record LoginRequest(
+        String personalId,
+        String password
+) {}
