+++ src/main/java/site/board/boardtraining/domain/auth/filter/JsonUsernamePasswordAuthenticationFilter.java
@@ -1,0 +1,97 @@
+package site.board.boardtraining.domain.auth.filter;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AuthenticationServiceException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.util.StreamUtils;
+import site.board.boardtraining.domain.auth.dto.LoginRequest;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+@Slf4j
+public class JsonUsernamePasswordAuthenticationFilter
+        extends AbstractAuthenticationProcessingFilter {
+
+    private static final String DEFAULT_LOGIN_REQUEST_URL = "/api/login";
+    private static final String HTTP_METHOD = "POST";
+    private static final String CONTENT_TYPE = "application/json";
+    private static final AntPathRequestMatcher DEFAULT_LOGIN_PATH_REQUEST_MATCHER =
+            new AntPathRequestMatcher(DEFAULT_LOGIN_REQUEST_URL, HTTP_METHOD);
+
+    private final ObjectMapper objectMapper;
+
+    public JsonUsernamePasswordAuthenticationFilter(
+            ObjectMapper objectMapper,
+            AuthenticationSuccessHandler authenticationSuccessHandler,
+            AuthenticationFailureHandler authenticationFailureHandler
+    ) {
+
+        super(DEFAULT_LOGIN_PATH_REQUEST_MATCHER);
+
+        this.objectMapper = objectMapper;
+        setAuthenticationSuccessHandler(authenticationSuccessHandler);
+        setAuthenticationFailureHandler(authenticationFailureHandler);
+    }
+
+    @Override
+    public Authentication attemptAuthentication(
+            HttpServletRequest request,
+            HttpServletResponse response
+    ) throws AuthenticationException, IOException, ServletException {
+
+        validateContentType(request);
+
+        LoginRequest loginRequest = parseLoginRequest(request);
+
+        validateLoginRequest(loginRequest);
+
+        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+                loginRequest.personalId(),
+                loginRequest.password()
+        );
+        setDetails(request, authenticationToken);
+
+        return this.getAuthenticationManager().authenticate(authenticationToken);
+    }
+
+    private void setDetails(
+            HttpServletRequest request,
+            UsernamePasswordAuthenticationToken authRequest
+    ) {
+        authRequest.setDetails(this.authenticationDetailsSource.buildDetails(request));
+    }
+
+    private void validateContentType(HttpServletRequest request) {
+        if (request.getContentType() == null || !request.getContentType().equals(CONTENT_TYPE)) {
+            throw new AuthenticationServiceException("Authentication Content-Type not supported: " + request.getContentType());
+        }
+    }
+
+    private LoginRequest parseLoginRequest(HttpServletRequest request)
+            throws IOException {
+        return objectMapper.readValue(
+                StreamUtils.copyToString(
+                        request.getInputStream(),
+                        StandardCharsets.UTF_8
+                ),
+                LoginRequest.class
+        );
+    }
+
+    private void validateLoginRequest(LoginRequest loginRequest) {
+        if (loginRequest.personalId() == null || loginRequest.password() == null) {
+            throw new AuthenticationServiceException("DATA IS MISS");
+        }
+    }
+}
