+++ src/main/java/site/board/boardtraining/domain/member/constant/MemberStatus.java
@@ -1,0 +1,5 @@
+package site.board.boardtraining.domain.member.constant;
+
+public enum MemberStatus {
+    ACTIVE, DELETE
+}
