+++ src/main/java/site/board/boardtraining/domain/member/dto/CreateMemberRequest.java
@@ -1,0 +1,22 @@
+package site.board.boardtraining.domain.member.dto;
+
+public record CreateMemberRequest(
+        String personalId,
+        String password,
+        String email,
+        String nickname
+) {
+    public static CreateMemberRequest of(
+            String personalId,
+            String password,
+            String email,
+            String nickname
+    ) {
+        return new CreateMemberRequest(
+                personalId,
+                password,
+                email,
+                nickname
+        );
+    }
+}
