+++ src/main/java/site/board/boardtraining/domain/member/dto/CreateMemberResponse.java
@@ -1,0 +1,11 @@
+package site.board.boardtraining.domain.member.dto;
+
+import site.board.boardtraining.domain.member.entity.Member;
+
+public record CreateMemberResponse(
+        Long memberId
+) {
+    public static CreateMemberResponse from(Member member) {
+        return new CreateMemberResponse(member.getId());
+    }
+}
