--- src/main/java/site/board/boardtraining/member/constant/MemberStatus.java
@@ -1,5 +1,0 @@
-package site.board.boardtraining.member.constant;
-
-public enum MemberStatus {
-    ACTIVE, DELETE
-}
